HnF Version: 1.0
HnFTabList is a simple plugin to have a Header and Footer on your player
TabList. Comes with colour codes (full list of colours in the config)
and a reload command for an easy Header or Footer config change update
without needing to restart the server! (players will need to relog to
see the change)